End-to-end test and GitHub Actions configuration
