+++ .idea/dictionaries/chris.xml
@@ -11,6 +11,7 @@
       <w>kmer</w>
       <w>rcount</w>
       <w>taattacggtcgtcgacaggcaggtagtaactcgcctatctgctgcaggctgcttagggtttcgtccgtg</w>
+      <w>transcriptome</w>
       <w>ttgcagcggatcaccagcaccaggtagtttcgtccgggtgtgaccgaaaggtaagagggagacccttgtc</w>
       <w>ttgcagcggatcaccagcaccaggtggtttcgtccgggtgtgaccgaaaggtaagagggagacccttgtc</w>
     </words>
