+++ testdata/lung_tissue_output
@@ -1,0 +1,57 @@
+ttagtgcaaagaatagagctcgcaccgt
+tgccattagtgcaaagaatagagctcgc
+gtgcaaagaatagagctcgcaccgtagc
+attagtgcaaagaatagagctcgcaccg
+atgccattagtgcaaagaatagagctcg
+agtgcaaagaatagagctcgcaccgtag
+tgcaaagaatagagctcgcaccgtagct
+tagtgcaaagaatagagctcgcaccgta
+tcaccttatgggttgggattatcctaaa
+gaaaaccctcaccttatgggttgggatt
+aaccctcaccttatgggttgggattatc
+aaaaccctcaccttatgggttgggatta
+tagaaaaccctcaccttatgggttggga
+ccctactataactcaaatgaatcttaag
+cattgtttgtagatttgacactagagtg
+atggtgtatgcctattttggaattgcaa
+agaaaaccctcaccttatgggttgggat
+aaaccctcaccttatgggttgggattat
+tgtctttagctatagatgcttacccact
+tggtgtatgcctattttggaattgcaat
+tttagctatagatgcttacccacttact
+ttcatttatgctttggtgtaaagatggc
+tctttagctatagatgcttacccactta
+gttctgatgttctttaccaaccaccaca
+gtctttagctatagatgcttacccactt
+ctttagctatagatgcttacccacttac
+cagctggttttccatttaataaatgggg
+agctggttttccatttaataaatggggt
+taaacatttctcaatgatgatactctct
+gtagtgctttattagaagatgaatttac
+ggttcatgtggtagtgttggttttaaca
+accgttattttaaatattgggatcagac
+tactatttgtagttgaagttgttgataa
+gttcatgtggtagtgttggttttaacat
+gtaaggctagactttattatgattcaat
+gtaaacatttctcaatgatgatactctc
+agtctactcttgtgcagaatgaattctc
+acttatgaaaactataggtccagacatg
+taaggctagactttattatgattcaatg
+gttggtaaacctagaccaccacttaacc
+gatgcacaatcgtttttaaacgggtttg
+ctaacaatgttgcttttcaaactgtcaa
+caatgttgcttttcaaactgtcaaaccc
+actatttgtagttgaagttgttgataag
+acaatgttgcttttcaaactgtcaaacc
+ttcaatggccacaaattatgatttgagt
+gttgcagttaaaatgtttgatgcttacg
+caaaaagagttgatttttgtggaaaggg
+gttgtgtacacacactggtactggtcag
+agttgcagttaaaatgtttgatgcttac
+gcacaggcactagtactgatgtcgtata
+cacaggcactagtactgatgtcgtatac
+aagattgctgattataattataaattac
+tactaaaatgtcagagtgtgtacttgga
+ctactaaaatgtcagagtgtgtacttgg
+gagctgcctatatggaagagccctaatg
+agagctgcctatatggaagagccctaat
